JdbcAvroOptions
