+++ dbeam-core/src/main/scala/com/spotify/dbeam/JdbcAvroJob.scala
@@ -19,7 +19,7 @@ package com.spotify.dbeam
 
 import java.sql.Connection
 
-import com.spotify.dbeam.options.{JdbcConnectionConfiguration, JdbcExportArgs}
+import com.spotify.dbeam.options.{JdbcAvroOptions, JdbcConnectionConfiguration, JdbcExportArgs}
 import org.apache.avro.Schema
 import org.apache.beam.sdk.metrics.Metrics
 import org.apache.beam.sdk.options.PipelineOptions
@@ -68,7 +68,7 @@ object JdbcAvroJob {
   def jdbcAvroTransform(output: String,
                         options: JdbcExportArgs,
                         generatedSchema: Schema): PTransform[PCollection[String], _ <: POutput] = {
-    val jdbcAvroOptions = JdbcAvroIO.JdbcAvroOptions.create(
+    val jdbcAvroOptions = JdbcAvroOptions.create(
       JdbcConnectionConfiguration.create(options.driverClass, options.connectionUrl)
         .withUsername(options.username)
         .withPassword(options.password),
